Initial Commit for the Assets folder
This commit included the initial prototype scene. you can currently use the 'f' key to make your player flip to the roof, or back down to the floor. there is one obstacle that if you hit, you will die. The current Player sprite, and spike sprite are not original images, and will need to be changed.