New Project
When I updated my unity, it broke the project file, so I made a new one, and re added all my scripts. I have made it so that if a block doesnt have another one to the left of it, it automatically gives itself spikes. working on have the player move properly. Currently he gets stuck, and Im not sure what the problem is at the moment